Updating reqs and schema as task, used django channels for async messaging

+++ requirements_manager/views.py
@@ -1,16 +1,18 @@
-from django.shortcuts import reverse, redirect
 from django.views.generic.list import ListView
 from django.views.generic import CreateView, UpdateView
 from django.contrib.auth.decorators import login_required
-from django.contrib import messages
 from django.http import JsonResponse
+from django.db import transaction
 
-from requirements_manager.models import Requirement
-from helpers.helper_mixins import ExperimentPackageTypeMixin
-from experiments_manager.helper import verify_and_get_experiment
 from git_manager.helpers.github_helper import GitHubHelper
-from requirements_manager.forms import RequirementForm
+from helpers.helper_mixins import ExperimentPackageTypeMixin
 from helpers.helper import get_package_or_experiment
+from experiments_manager.consumers import send_message
+
+from .forms import RequirementForm
+from .models import Requirement
+from .mixins import RequirementSuccessUrlMixin
+from .tasks import task_write_requirements_file
 
 
 class RequirementListView(ExperimentPackageTypeMixin, ListView):
@@ -28,7 +30,7 @@ class RequirementListView(ExperimentPackageTypeMixin, ListView):
         return context
 
 
-class RequirementCreateView(ExperimentPackageTypeMixin, CreateView):
+class RequirementCreateView(ExperimentPackageTypeMixin, RequirementSuccessUrlMixin, CreateView):
     model = Requirement
     fields = ['package_name', 'version']
     template_name = 'requirements_manager/requirement_list.html'
@@ -45,16 +47,8 @@ class RequirementCreateView(ExperimentPackageTypeMixin, CreateView):
         obj.requirements.add(req)
         obj.save()
 
-    def get_success_url(self):
-        if self.kwargs['object_type'] == self.EXPERIMENT_TYPE:
-            experiment = verify_and_get_experiment(self.request, self.kwargs['object_id'])
-            success_url = reverse('experiment_detail', kwargs={'pk': experiment.id, 'slug': experiment.slug()})
-        elif self.kwargs['object_type']  == self.PACKAGE_TYPE:
-            success_url = reverse('internalpackage_dashboard', kwargs={'pk': self.kwargs['object_id']})
-        return success_url
 
-
-class RequirementUpdateView(ExperimentPackageTypeMixin, UpdateView):
+class RequirementUpdateView(ExperimentPackageTypeMixin, RequirementSuccessUrlMixin, UpdateView):
     model = Requirement
     fields = ['package_name', 'version']
 
@@ -72,26 +66,23 @@ def remove_experiment_requirement(request, object_id, object_type):
         assert 'requirement_id' in request.POST
         requirement_id = request.POST['requirement_id']
         exp_or_package = get_package_or_experiment(request, object_type, object_id)
-        requirement = Requirement.objects.get(id=requirement_id)
-        exp_or_package.requirements.remove(requirement)
-        exp_or_package.save()
-        requirement.delete()
-        return JsonResponse({'deleted': True})
+        requirement = exp_or_package.requirements.filter(pk=requirement_id)
+        if requirement:
+            requirement = requirement[0]
+            with transaction.atomic():
+                exp_or_package.requirements.remove(requirement)
+                exp_or_package.save()
+                requirement.delete()
+            return JsonResponse({'deleted': True})
+    return JsonResponse({'deleted': False})
 
 
 @login_required
 def write_requirements_file(request, object_id, object_type):
-    exp_or_package = get_package_or_experiment(request, object_type, object_id)
-    requirements_txt = build_requirements_file(exp_or_package)
-    github_helper = GitHubHelper(request.user, exp_or_package.git_repo.name)
-    github_helper.update_file('requirements.txt', 'Updated requirements.txt file by MOOC workbench',
-                              requirements_txt)
-    messages.add_message(request, messages.INFO, 'Successfully updated requirements in your repository')
-    return redirect(to=exp_or_package.get_absolute_url())
+    get_package_or_experiment(request, object_type, object_id)
+    task_write_requirements_file.delay(object_id, object_type)
+    send_message(request.user.username, 'info', 'Task started to update dependencies...')
+    return JsonResponse({'success': True})
 
 
-def build_requirements_file(exp_or_package):
-    requirements_txt = ''
-    for requirement in exp_or_package.requirements.all():
-        requirements_txt += '{0}\n'.format(str(requirement))
-    return requirements_txt
+
